+++ src/syntax/parsers/expression/mod.rs
@@ -9,6 +9,7 @@ mod primary;
 mod term;
 mod unary;
 mod utils;
+mod array;
 
 impl<'a> Parseable<'a> for Expression<'a> {
     type Item = Expression<'a>;